requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
comtypes==1.1.10
idna==3.3
MouseInfo==0.1.3
Pillow==9.0.1
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.28
pyttsx3==2.71
pytweening==1.0.4
pywhatkit==5.1
pywin32==302
requests==2.26.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
urllib3==1.26.7
wikipedia==1.4.0
